5_django_app_mgmt/requirements.txt
Django>4,<5
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-pdb==0.6.2
docutils==0.18.1
raven==6.10.0
